requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes --no-editable --no-emit-project --output-file requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   pytest
    #   tqdm
    #   uvicorn
confluent-kafka==2.3.0
    # via finder-enrichment
deprecated==1.2.18
    # via limits
fastapi==0.116.1
    # via
    #   finder-enrichment
    #   finder-enrichment-db-client
    #   finder-enrichment-db-contracts
finder-enrichment-db-client==0.1.5
    # via finder-enrichment
finder-enrichment-db-contracts==0.1.5
    # via
    #   finder-enrichment
    #   finder-enrichment-db-client
google-ai-generativelanguage==0.6.10
    # via google-generativeai
google-api-core==2.25.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.179.0
    # via google-generativeai
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.3
    # via finder-enrichment
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.74.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   finder-enrichment-db-client
    #   listings-db-api-client
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
kafka-python==2.0.2
    # via finder-enrichment
limits==5.5.0
    # via slowapi
listings-db-api-client==1.0.5
    # via finder-enrichment
listings-db-contracts==1.0.4
    # via
    #   finder-enrichment
    #   listings-db-api-client
numpy==2.3.2
    # via pandas
packaging==25.0
    # via
    #   limits
    #   pytest
pandas==2.3.2
    # via finder-enrichment
pluggy==1.6.0
    # via pytest
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   fastapi
    #   finder-enrichment-db-client
    #   finder-enrichment-db-contracts
    #   google-generativeai
    #   listings-db-contracts
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via pytest
pyparsing==3.2.3
    # via httplib2
pytest==8.4.1
    # via finder-enrichment
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via
    #   finder-enrichment
    #   listings-db-api-client
    #   uvicorn
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via uvicorn
requests==2.32.5
    # via google-api-core
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
slowapi==0.1.9
    # via finder-enrichment
sniffio==1.3.1
    # via anyio
starlette==0.47.2
    # via fastapi
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.14.1
    # via
    #   anyio
    #   fastapi
    #   google-generativeai
    #   limits
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via finder-enrichment
uvloop==0.21.0 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wrapt==1.17.3
    # via deprecated
